PluginSettings.hpp: sync with Far PluginSettings.hpp

+++ src/PluginSDK/Far3/PluginSettings.hpp
@@ -25,34 +25,34 @@ public:
     SettingsControl(handle,SCTL_FREE,0,0);
   }
 
-  intptr_t CreateSubKey(intptr_t Root, const wchar_t *Name)
+  int CreateSubKey(size_t Root, const wchar_t *Name)
   {
     FarSettingsValue value={sizeof(FarSettingsValue),Root,Name};
-    return SettingsControl(handle,SCTL_CREATESUBKEY,0,&value);
+    return (int)SettingsControl(handle,SCTL_CREATESUBKEY,0,&value);
   }
 
-  intptr_t OpenSubKey(intptr_t Root, const wchar_t *Name)
+  int OpenSubKey(size_t Root, const wchar_t *Name)
   {
     // DEBUG_PRINTF(L"handle = %d, Name = %s", (int)handle, Name);
     FarSettingsValue value={sizeof(FarSettingsValue),Root,Name};
-    intptr_t Result = SettingsControl(handle,SCTL_OPENSUBKEY,0,&value);
+    int Result = (int)SettingsControl(handle,SCTL_OPENSUBKEY,0,&value);
     // DEBUG_PRINTF(L"Result = %d", Result);
     return Result;
   }
 
-  bool DeleteSubKey(int Root)
+  bool DeleteSubKey(size_t Root)
   {
     FarSettingsValue value={sizeof(FarSettingsValue),Root,nullptr};
     return SettingsControl(handle,SCTL_DELETE,0,&value) ? true : false;
   }
 
-  bool DeleteValue(int Root, const wchar_t *Name)
+  bool DeleteValue(size_t Root, const wchar_t *Name)
   {
     FarSettingsValue value={sizeof(FarSettingsValue),Root,Name};
     return SettingsControl(handle,SCTL_DELETE,0,&value) ? true : false;
   }
 
-  const wchar_t *Get(int Root, const wchar_t *Name, const wchar_t *Default)
+  const wchar_t *Get(size_t Root, const wchar_t *Name, const wchar_t *Default)
   {
     FarSettingsItem item={sizeof(FarSettingsItem),Root,Name,FST_STRING};
     if (SettingsControl(handle,SCTL_GET,0,&item))
@@ -62,12 +62,12 @@ public:
     return Default;
   }
 
-  void Get(int Root, const wchar_t *Name, wchar_t *Value, size_t Size, const wchar_t *Default)
+  void Get(size_t Root, const wchar_t *Name, wchar_t *Value, size_t Size, const wchar_t *Default)
   {
     lstrcpyn(Value, Get(Root,Name,Default), (int)Size);
   }
 
-  unsigned __int64 Get(int Root, const wchar_t *Name, unsigned __int64 Default)
+  unsigned __int64 Get(size_t Root, const wchar_t *Name, unsigned __int64 Default)
   {
     FarSettingsItem item={sizeof(FarSettingsItem),Root,Name,FST_QWORD};
     if (SettingsControl(handle,SCTL_GET,0,&item))
@@ -77,13 +77,13 @@ public:
     return Default;
   }
 
-  __int64      Get(int Root, const wchar_t *Name, __int64 Default) { return (__int64)Get(Root,Name,(unsigned __int64)Default); }
-  int          Get(int Root, const wchar_t *Name, int Default)  { return (int)Get(Root,Name,(unsigned __int64)Default); }
-  unsigned int Get(int Root, const wchar_t *Name, unsigned int Default) { return (unsigned int)Get(Root,Name,(unsigned __int64)Default); }
-  DWORD        Get(int Root, const wchar_t *Name, DWORD Default) { return (DWORD)Get(Root,Name,(unsigned __int64)Default); }
-  bool         Get(int Root, const wchar_t *Name, bool Default) { return Get(Root,Name,Default?1ull:0ull)?true:false; }
+  __int64      Get(size_t Root, const wchar_t *Name, __int64 Default) { return (__int64)Get(Root,Name,(unsigned __int64)Default); }
+  int          Get(size_t Root, const wchar_t *Name, int Default)  { return (int)Get(Root,Name,(unsigned __int64)Default); }
+  unsigned int Get(size_t Root, const wchar_t *Name, unsigned int Default) { return (unsigned int)Get(Root,Name,(unsigned __int64)Default); }
+  DWORD        Get(size_t Root, const wchar_t *Name, DWORD Default) { return (DWORD)Get(Root,Name,(unsigned __int64)Default); }
+  bool         Get(size_t Root, const wchar_t *Name, bool Default) { return Get(Root,Name,Default?1ull:0ull)?true:false; }
 
-  size_t Get(int Root, const wchar_t *Name, void *Value, size_t Size)
+  size_t Get(size_t Root, const wchar_t *Name, void *Value, size_t Size)
   {
     FarSettingsItem item={sizeof(FarSettingsItem),Root,Name,FST_DATA};
     if (SettingsControl(handle,SCTL_GET,0,&item))
@@ -95,27 +95,27 @@ public:
     return 0;
   }
 
-  bool Set(int Root, const wchar_t *Name, const wchar_t *Value)
+  bool Set(size_t Root, const wchar_t *Name, const wchar_t *Value)
   {
     FarSettingsItem item={sizeof(FarSettingsItem),Root,Name,FST_STRING};
     item.String=Value;
     return SettingsControl(handle,SCTL_SET,0,&item)!=FALSE;
   }
 
-  bool Set(int Root, const wchar_t *Name, unsigned __int64 Value)
+  bool Set(size_t Root, const wchar_t *Name, unsigned __int64 Value)
   {
     FarSettingsItem item={sizeof(FarSettingsItem),Root,Name,FST_QWORD};
     item.Number=Value;
     return SettingsControl(handle,SCTL_SET,0,&item)!=FALSE;
   }
 
-  bool Set(int Root, const wchar_t *Name, __int64 Value) { return Set(Root,Name,(unsigned __int64)Value); }
-  bool Set(int Root, const wchar_t *Name, int Value) { return Set(Root,Name,(unsigned __int64)Value); }
-  bool Set(int Root, const wchar_t *Name, unsigned int Value) { return Set(Root,Name,(unsigned __int64)Value); }
-  bool Set(int Root, const wchar_t *Name, DWORD Value) { return Set(Root,Name,(unsigned __int64)Value); }
-  bool Set(int Root, const wchar_t *Name, bool Value) { return Set(Root,Name,Value?1ull:0ull); }
+  bool Set(size_t Root, const wchar_t *Name, __int64 Value) { return Set(Root,Name,(unsigned __int64)Value); }
+  bool Set(size_t Root, const wchar_t *Name, int Value) { return Set(Root,Name,(unsigned __int64)Value); }
+  bool Set(size_t Root, const wchar_t *Name, unsigned int Value) { return Set(Root,Name,(unsigned __int64)Value); }
+  bool Set(size_t Root, const wchar_t *Name, DWORD Value) { return Set(Root,Name,(unsigned __int64)Value); }
+  bool Set(size_t Root, const wchar_t *Name, bool Value) { return Set(Root,Name,Value?1ull:0ull); }
 
-  bool Set(int Root, const wchar_t *Name, const void *Value, size_t Size)
+  bool Set(size_t Root, const wchar_t *Name, const void *Value, size_t Size)
   {
     FarSettingsItem item={sizeof(FarSettingsItem),Root,Name,FST_DATA};
     item.Data.Size=Size;
@@ -125,19 +125,21 @@ public:
 
   bool GetSubKeyNames(FarSettingsEnum &settings)
   {
-        return SettingsControl(handle, SCTL_ENUM, 0, &settings) != FALSE;
+    return SettingsControl(handle, SCTL_ENUM, 0, &settings) != FALSE;
   }
-  bool ValueExists(int Root, const wchar_t *Name)
+
+  bool ValueExists(size_t Root, const wchar_t *Name)
   {
     FarSettingsItem item = {sizeof(FarSettingsItem), Root, Name, FST_DATA};
     return SettingsControl(handle, SCTL_GET, 0, &item) != FALSE;
   }
-  int BinaryDataSize(int Root, const wchar_t *Name)
+
+  int BinaryDataSize(size_t Root, const wchar_t *Name)
   {
     FarSettingsItem item = {sizeof(FarSettingsItem), Root, Name, FST_DATA};
     if (SettingsControl(handle, SCTL_GET, 0, &item) != FALSE)
     {
-       return item.Data.Size;
+       return (int)item.Data.Size;
     }
     return 0;
   }
